--- a/setup.py
+++ b/setup.py
@@ -90,12 +90,8 @@ params = dict(
         'scp', # SSH connection file transfers
         'wrapt',  # Basic for construction of decorator functions
         'future', # Python 2-3 compatibility
-        'enum34;python_version<"3.4"', # Enums for Python < 3.4
-        'contextlib2;python_version<"3.0"', # Python 3 contextlib backport for Python 2
-        'numpy<=1.16.4; python_version<"3"',
-        'numpy; python_version>="3"',
-        'pandas<=0.24.2; python_version<"3"',
-        'pandas; python_version>"3"',
+        'numpy',
+        'pandas',
         'lxml', # More robust xml parsing
     ],
     extras_require={