@douglas-raillard-arm 2021-08-19
Preliminary results on:
```
        with target.cpufreq.use_governor('performance'):
            pass
```

30% speedup On a SSH target (juno r0):
    * on this branch: 3.7s (avg over 10 runs)
    * on master:        5.3s (avg over 10 runs)

50% speedup on a local target:
    * on this branch: ~9s 
    * on master:        ~20s