@douglas-raillard-arm 2021-08-31
There are a few things to cleanup (docstring etc) so if you are happy with the API as it stands I will proceed with:

    * adding docstrings
    * renaming asyn.parallel into asyn.concurrently
    * remove the .concurrent attribute, as it is not that useful in practice I think (it allows running the function concurrently, only blocking to get the result when it is awaited. The "normal" couroutine behavior is to not do anything until the return value is awaited). asyn.concurrently is usually better as it makes it obvious what runs concurrently and avoids any kind of "task leak".