@douglas-raillard-arm 2021-11-11
Updated the PR with checks to make sure that no file access conflicts when executing concurrent asyncio tasks. An artificial example:
```
from devlib.utils.asyn import *

m = ConcurrentResourceManager()

r1 = FileResource('host', 'file1', 'w')
r2 = FileResource('host', 'file1', 'r')

async def f1():
    # m.track_resource(r1)
    m.track_resource(r2)
    print('from f1')


async def f2():
    m.track_resource(r1)
    m.track_resource(r2)
    print('from f2')

async def f3():
    # m.track_resource(r1)
    print('from f3')

run(
    m.concurrently(
        (
            f1(),
            m.concurrently(
                (
                    f2(),
                    f3(),
                )
            )
        )
    )
)
```
In the PR, `ConcurrentResourceManager` is `target.resource_manager`, which is thread-local