@douglas-raillard-arm 2021-11-15
PR updated with:
* forcing python 3.7 in setup.py
* Rename `ConcurrentResourceManager` to `AsyncManager`
* Almost all methods of `Target` have been converted. Not all of them have been "smartly" converted though, e.g. by doing things concurrently. 
* Some more modules/collector updated to make use of concurrent commands (cgroups module init)