@douglas-raillard-arm 2021-11-23
@marcbonnici @setrofim 
Updated with a bunch of fixes that were needed to run on my test platform:
* Add `Target.execute(..., decode=True)` parameter. When `decode=False` is used, execute() returns  a tuple of bytes (stdout, stderr). This was necessary to successfully read sysfs content, which can contain binary data (from device tree). Note that the change to the Android connection class is entirely untested so a simple `Target.execute('echo hello; echo world >&2 ', decode=False)` on an android device would be useful.

* Cleanup the cgroup module. Trying to parallelize the setup shed some light on a few problems with the code structure. This PR migrates it to use cgroup v2, cleans some cruft and exposes more clearly what pieces will be retainable in a future cgroup API that allows full cgroup power and what will probably become deprecated.