@douglas-raillard-arm 2021-11-24
@marcbonnici PR updated with:

fixes to cgroups:
* only manipulate PIDs since thread support in cgroup v2 is significantly more complex and would require a dedicated API. This is probably best left for a future API
* Fix to `Controller.tasks('/')`. Since existing code apparently expects to list all the tasks in the root group (rather than the root of the subtree handled by the controller), handle it as a special case.

Fix to Target.install():
Add a `has_busybox=True` parameter to Connection.execute(), so that install() is able to install busybox without needing to run it (for Android). This is the WIP commit which will be squashed if it works well.