@douglas-raillard-arm 2021-11-26
@marcbonnici @setrofim Updated the PR:
* turned Target.execute(decode=False) into a separate method `Target.execute_raw()`
* Implemented `Target.execute_raw()` on top of BackgroundCommand.communicate() to avoid invasive changes to Connection.execute().
* removed cgroups modification, as they would be better served by a new module
* Added `busybox --install -s` and the corresponding simplification in the code