@douglas-raillard-arm 2022-01-24
I've updated the PR with:
* fixes for missing () and await
* An (very unlikely) possible fix for the bg command benchmarking hang:
```
-            '{} xargs kill -9'.format(self.busybox),
+            '{} xargs kill -9 --'.format(self.busybox),
```