@douglas-raillard-arm 2022-02-02
Updated PR. 
Added:
* Added `--` to `devlib.connection._kill_pgid_cmd` to avoid issues as well

Still to do:
* split the addition of busybox to PATH in a separate PR (and fix the issues with ps output)
* figure out why `bg.__exit__` is hanging on your setup