@marcbonnici 2022-02-22
Finally found the issue on my android device and it was due to the PID reporting of adb connections. 
Occasionally it would report the PID of the command used to find the PID rather than the PID of the command itself due to PID reuse. 
I've submitted a [PR  ](https://github.com/ARM-software/devlib/pull/581)to fix this so can continue with on my android devices.