@douglas-raillard-arm 2022-04-01
PR updated:

Changes:
* Added max_bg connection parameter, so that users can force serialized execute() and also limit the auto-detection duration.
* split the PR for busybox symlinks so it's not blocking this one: https://github.com/ARM-software/devlib/pull/585

TODO:
* workaround/fix the android background command PID detection problem