@douglas-raillard-arm 2022-04-01
PR updated again with:
* fix the android background command PID detection problem. 
This is somewhat redundant with https://github.com/ARM-software/devlib/pull/581 but tries another approach. Instead of printing the PID to stdout, it writes it to a temp file. This avoids the issues of streams redirection, that are acceptable for SSH since it needs to be done anyway, but for adb it would add a great deal of complexity.